adds bifrost ui
