+++ transports/bifrost-http/test-simple.json
@@ -1,0 +1,160 @@
+{
+  "providers": {
+    "openai": {
+      "keys": [
+        {
+          "value": "env.OPENAI_API_KEY",
+          "models": [
+            "gpt-3.5-turbo",
+            "gpt-3.5-turbo-preview",
+            "gpt-4",
+            "gpt-4o",
+            "gpt-4o-mini",
+            "gpt-4-turbo",
+            "gpt-4-turbo-preview",
+            "gpt-4-vision-preview"
+          ],
+          "weight": 1.0
+        }
+      ],
+      "network_config": {
+        "extra_headers": {
+          "X-Organization-ID": "org-123",
+          "X-Environment": "production"
+        },
+        "default_request_timeout_in_seconds": 30,
+        "max_retries": 1,
+        "retry_backoff_initial_ms": 100,
+        "retry_backoff_max_ms": 2000
+      },
+      "concurrency_and_buffer_size": {
+        "concurrency": 3,
+        "buffer_size": 10
+      }
+    },
+    "anthropic": {
+      "keys": [
+        {
+          "value": "env.ANTHROPIC_API_KEY",
+          "models": [
+            "claude-2.1",
+            "claude-3-sonnet-20240229",
+            "claude-3-haiku-20240307",
+            "claude-3-opus-20240229",
+            "claude-3-5-sonnet-20240620",
+            "claude-3-7-sonnet-20250219"
+          ],
+          "weight": 1.0
+        }
+      ],
+      "network_config": {
+        "default_request_timeout_in_seconds": 30,
+        "max_retries": 1,
+        "retry_backoff_initial_ms": 100,
+        "retry_backoff_max_ms": 2000
+      },
+      "concurrency_and_buffer_size": {
+        "concurrency": 3,
+        "buffer_size": 10
+      }
+    },
+    "bedrock": {
+      "keys": [
+        {
+          "value": "env.BEDROCK_API_KEY",
+          "models": [
+            "anthropic.claude-v2:1",
+            "mistral.mixtral-8x7b-instruct-v0:1",
+            "mistral.mistral-large-2402-v1:0",
+            "anthropic.claude-3-sonnet-20240229-v1:0"
+          ],
+          "weight": 1.0
+        }
+      ],
+      "network_config": {
+        "default_request_timeout_in_seconds": 30,
+        "max_retries": 1,
+        "retry_backoff_initial_ms": 100,
+        "retry_backoff_max_ms": 2000
+      },
+      "meta_config": {
+        "secret_access_key": "env.AWS_SECRET_ACCESS_KEY",
+        "region": "us-east-1"
+      },
+      "concurrency_and_buffer_size": {
+        "concurrency": 3,
+        "buffer_size": 10
+      }
+    },
+    "cohere": {
+      "keys": [
+        {
+          "value": "env.COHERE_API_KEY",
+          "models": ["command-a-03-2025"],
+          "weight": 1.0
+        }
+      ],
+      "network_config": {
+        "default_request_timeout_in_seconds": 30,
+        "max_retries": 1,
+        "retry_backoff_initial_ms": 100,
+        "retry_backoff_max_ms": 2000
+      },
+      "concurrency_and_buffer_size": {
+        "concurrency": 3,
+        "buffer_size": 10
+      }
+    },
+    "azure": {
+      "keys": [
+        {
+          "value": "env.AZURE_API_KEY",
+          "models": ["gpt-4o"],
+          "weight": 1.0
+        }
+      ],
+      "network_config": {
+        "default_request_timeout_in_seconds": 30,
+        "max_retries": 1,
+        "retry_backoff_initial_ms": 100,
+        "retry_backoff_max_ms": 2000
+      },
+      "meta_config": {
+        "endpoint": "env.AZURE_ENDPOINT",
+        "deployments": {
+          "gpt-4o": "gpt-4o-aug"
+        },
+        "api_version": "2024-08-01-preview"
+      },
+      "concurrency_and_buffer_size": {
+        "concurrency": 3,
+        "buffer_size": 10
+      }
+    },
+    "vertex": {
+      "keys": [],
+      "meta_config": {
+        "project_id": "env.VERTEX_PROJECT_ID",
+        "region": "us-central1",
+        "auth_credentials": "env.VERTEX_CREDENTIALS"
+      },
+      "concurrency_and_buffer_size": {
+        "concurrency": 3,
+        "buffer_size": 10
+      }
+    }
+  },
+  "mcp": {
+    "client_configs": [
+      {
+        "name": "your-mcp-server-name",
+        "connection_type": "stdio",
+        "stdio_config": {
+          "command": "npx",
+          "args": ["-y", "your-mcp-server-name"],
+          "envs": ["YOUR_MCP_SERVER_ENV_VAR"]
+        }
+      }
+    ]
+  }
+}
