--- transports/bifrost-http/tracking/prometheus.yml
@@ -1,15 +1,0 @@
-# Prometheus configuration for tracking bifrost-http service (for development and testing purposes only, don't use in production without proper setup)
-global:
-  scrape_interval: 5s # Scrape every 5 seconds
-
-# Note: Target configuration depends on your deployment environment:
-# - For local development: Use "host.docker.internal:8080" to access the service running on your host machine
-# - For Docker deployment: Use "bifrost-api:8080" to access the service within the Docker network
-#   Make sure to replace "bifrost-api" and "8080" with your actual docker container name and port if different
-#   Also check that you have the bifrost container inside "bifrost_tracking_network".
-
-scrape_configs:
-  - job_name: "bifrost-api"
-    static_configs:
-      - targets: ["host.docker.internal:8080"] # Scrape from the /metrics endpoint
-
