+++ ui/README.md
@@ -1,0 +1,15 @@
+# Bifrost UI
+
+Web interface to manage the Bifrost AI provider gateway.
+
+
+## Development
+
+- `npm run dev` - Start development server
+- `npm run build` - Build for production
+- `npm run start` - Start production server
+- `npm run lint` - Run ESLint
+
+## Integration with Bifrost HTTP
+
+The static build output can be served by the Bifrost HTTP transport server. Place the contents of the `out` directory in your web server's static file directory.
