+++ ui/app/layout.tsx
@@ -1,0 +1,30 @@
+import { Sidebar } from '@/components/sidebar'
+import { Inter } from 'next/font/google'
+import { ReactNode } from 'react'
+import './globals.css'
+
+const inter = Inter({ subsets: ['latin'] })
+
+export const metadata = {
+  title: 'Bifrost UI',
+  description: 'Configuration interface for Bifrost AI provider gateway'
+}
+
+interface RootLayoutProps {
+  children: ReactNode
+}
+
+export default function RootLayout({ children }: RootLayoutProps) {
+  return (
+    <html lang="en" className="dark">
+      <body className={`${inter.className} antialiased`}>
+        <div className="flex h-screen bg-background">
+          <Sidebar />
+          <main className="flex-1 overflow-auto">
+            {children}
+          </main>
+        </div>
+      </body>
+    </html>
+  )
+}
